str-monitor/scraper/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
pdfplumber==0.11.4
supabase==2.10.0
resend==2.5.0
python-dotenv==1.0.1
lxml==5.3.0
